Set up the DI startup for the endpoint. Also renamed ctrl to handler to avoid confusion.
